requirements.txt
sentence-transformers~=2.7.0
python-dotenv~=1.0.1
Markdown~=3.6
beautifulsoup4~=4.12.3 # Optional: For cleaner text extraction from Markdown->HTML
nltk~=3.8.1 # For sentence tokenization (better chunking)
langchain~=0.3.25 # Or a newer version
langchain-community~=0.3.23 # Or a newer version
langchain-core~=0.3.58 # Or a newer version
langchain-ollama~=0.3.2
unstructured[md]~=0.17.2
numpy~=2.2.5
faiss-cpu~=1.11.0
